+++ qcfractal/interface/__init__.py
@@ -2,11 +2,11 @@
 DQM Client base folder
 """
 
+from . import collections
 from . import data
 from . import dict_utils
 from . import orm
 from . import schema
 from .client import FractalClient
-from . import collections
 # Add imports here
 from .molecule import Molecule