+++ qcfractal/interface/tests/test_database.py
@@ -2,9 +2,10 @@
 Tests the QCPortal database object
 """
 
+import pytest
+
 from . import portal
 from . import test_helper as th
-import pytest
 
 
 def _compare_stoichs(stoich, stoich_other):