+++ qcfractal/web_handlers.py
@@ -2,6 +2,7 @@
 Web handlers for the FractalServer
 """
 import json
+
 import tornado.web